Moves this from pvp tools to Loot Tracker
