+++ runelite-client/src/main/java/net/runelite/client/plugins/pvptools/PvpToolsConfig.java
@@ -207,15 +207,4 @@ public interface PvpToolsConfig extends Config
 		return true;
 	}
 
-	@ConfigItem(
-		keyName = "lootValueMessage",
-		name = "Loot Value Messages",
-		description = "Sends a game message with the total value you of your loot when you get a kill",
-		position = 16
-	)
-	default boolean sendLootValueMessages()
-	{
-		return true;
-	}
-
 }
